+++ src/main/java/io/bookbar/bookbarbackend/repository/BooksRepository.java
@@ -3,5 +3,5 @@ package io.bookbar.bookbarbackend.repository;
 import io.bookbar.bookbarbackend.entities.Books;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BooksRepository extends JpaRepository<Books, Long> {
+public interface BooksRepository extends JpaRepository<Books, Integer> {
 }
